requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0
python-multipart==0.0.6

# Web Scraping
playwright==1.41.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.2
lxml==5.1.0
selectorlib==0.16.0
requests==2.31.0
fake-useragent==1.4.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
sqlite-utils==3.36

# Location & ZIP Code
uszipcode==1.0.1
geopy==2.4.1

# APIs
ebay-rest-api==0.0.1
python-ebay-sdk==1.0.0

# Utilities
httpx==0.26.0
aiohttp==3.9.1
asyncio==3.4.3
tenacity==8.2.3
ratelimit==2.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
